feat(SimuNEX): automate variable initialization.
