+++ UnityProject/Assets/Scripts/Core/ObjectExtensions.cs.meta
